feat: adding swagger stats to project
