+++ package.json
@@ -49,6 +49,7 @@
   },
   "dependencies": {
     "@types/express-rate-limit": "^5.1.3",
+    "@types/swagger-stats": "^0.95.4",
     "axios": "^0.21.1",
     "config": "^3.3.6",
     "dotenv-cli": "^4.0.0",
@@ -59,6 +60,7 @@
     "joi": "^17.4.2",
     "module-alias": "^2.2.2",
     "pino": "^6.13.0",
+    "swagger-stats": "^0.99.2",
     "swagger-ui-express": "^4.1.6",
     "testlink-xmlrpc": "^2.0.2"
   }
